refactor: consolidate shared types into centralized package
